requirements.txt
# Enhanced requirements.txt for cloud deployment
streamlit>=1.28.0
pdf2image>=3.0.0
pytesseract>=0.3.10
sentence-transformers>=2.2.0
openai>=1.0.0
faiss-cpu>=1.7.0
Pillow>=9.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
requests>=2.31.0